Add View Controller with Scroll View to show the steps in order
It showed the steps in an alert, and only showed the results of each
step. It needed to show the calculations that were made to get to the
next step. In the results there were unnecessary steps taken, from the
selected unit to the base unit and then to the desired unit.

* Add user can dismiss keyboard tapping outside the keyboard
* Add close button to the steps view
* Add a step number title to the step nib
* Add a line below the title of the steps to show separation from steps
* Add unit conversion label in steps view to show which units are converting
* Clean and refactor Conversion class
* Add a conversion factor to each step in Step View Controller
* Change arrow image to vector (pdf)